Auto generate handlers using safe/unsafe/noop options
