+++ vulnerabilities/sqlInjection/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const shared = require('../shared');
+
+module.exports = shared('sqlInjection');
