+++ vulnerabilities/sqlInjection/views/index.ejs
@@ -1,3 +1,3 @@
 <% include ../../../views/partials/header %>
-<% include ../../../node_modules/@contrast/test-bench-content/views/sql-injection.ejs %>
+<% include ../../../node_modules/@contrast/test-bench-content/views/sqlInjection.ejs %>
 <% include ../../../views/partials/footer %>
